% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bentQRgrid.R
\name{bentQRgrid}
\alias{bentQRgrid}
\title{Grid search algorithm for bent line quantile regression}
\usage{
bentQRgrid(y, x, z, tau)
}
\arguments{
\item{y}{A vector of response}

\item{x}{A numeric variable with change point}

\item{z}{A vector of covariates}

\item{tau}{quantile level}
}
\value{
A list with the elements
\item{bet.est}{The estimated regression coefficients with intercept.}
\item{bp.est}{The estimated change point.}
\item{bet.se}{The estimated standard error of the regression coefficients.}
\item{bp.se}{The estimated standard error of the change point.}
}
\description{
This function estimates the bent line regression model
by grid search algorithm.
}
\examples{
\dontrun{

## simulated data
n <- 200
t0 <- 1.5    # true change point
bet0 <- c(1, 3, -2, 1)     # true regression coefficients
tau <- 0.7
modtype <- 1
etype <- 1
dat <- bentQRsimdat(n, bet0, t0, tau, modtype, etype)
head(dat) 
y <- dat$y
x <- dat$x
z <- dat$z

bentQRgrid(y, x, z, tau)

## MRS data
data(data_mrs)
x <- log(data_mrs$mass)
y <- log(data_mrs$speed)
z <- data_mrs$hopper
tau <- 0.5
bentQRgrid(y, x, z, tau)
}
}
\author{
Feipeng Zhang
}
\references{
Li, C., Wei, Y., Chappell, R., and He, X. (2011). 
Bent line quantile regression with application to
an allometric study of land mammals' speed and mass.Biometrics, 67:242--249.
}
\keyword{bentQRgrid}

