% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bentQRseg.R
\name{bentQRseg}
\alias{bentQRseg}
\title{Segmented estimate for bent line quantile regression}
\usage{
bentQRseg(y, x, z, tau, bet.ini, bp.ini, modtype, tol = 1e-04,
  max.iter = 100)
}
\arguments{
\item{y}{A vector of response}

\item{x}{A numeric variable with change point}

\item{z}{A vector of covariates}

\item{tau}{quantile level}

\item{bet.ini}{A initial vector of regression coefficients}

\item{bp.ini}{A initial value of change point}

\item{modtype}{type of error, modtype =1 for iid, modtype = 2 for nid}

\item{tol}{tolerance value, 1e-4 for default}

\item{max.iter}{the maximum iteration steps, 100 for default}
}
\value{
A list with the elements
\item{bet.est}{The estimated regression coefficients with intercept.}
\item{bp.est}{The estimated change point.}
\item{bet.se}{The estimated standard error of the regression coefficients.}
\item{bp.se}{The estimated standard error of the change point.}
\item{iter}{The iteration steps.}
}
\description{
This function estimates the bent line quantile regression model
by segmented estimation.
}
\examples{

\dontrun{

## simulated data
n <- 200
t0 <- 1.5    # true change point
bet0 <- c(1, 3, -2, 1)     # true regression coefficients
tau <- 0.7
modtype <- 1
etype <- 1
dat <- bentQRsimdat(n, bet0, t0, tau, modtype, etype)
head(dat) 
y <- dat$y
x <- dat$x
z <- dat$z

fit.grid <- bentQRgrid(y, x, z, tau)
bet.ini <- fit.grid$bet.est
bp.ini <- fit.grid$bp.est
bentQRseg(y, x, z, tau, bet.ini, bp.ini, modtype = 1)

## MRS data
data(data_mrs)
x <- log(data_mrs$mass)
y <- log(data_mrs$speed)
z <- data_mrs$hopper
tau <- 0.5

fit.grid <- bentQRgrid(y, x, z, tau)
bet.ini <- fit.grid$bet.est
bp.ini <- fit.grid$bp.est
bentQRseg(y, x, z, tau, bet.ini, bp.ini, modtype = 1)
}
}
\author{
Feipeng Zhang
}
\seealso{
\code{bentQRgrid}
}
\keyword{bentQRseg}

