% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bentQRsimdat.R
\name{bentQRsimdat}
\alias{bentQRsimdat}
\title{Simulated data from the bent line quantile regression}
\usage{
bentQRsimdat(n, bet0, t0, tau, modtype, etype)
}
\arguments{
\item{n}{sample size.}

\item{bet0}{the vecotr of true regression coefficients.}

\item{t0}{the true location of the change point.}

\item{tau}{the quantile level.}

\item{modtype}{type of model, where
modtype = 1, IID case,
modetype = 2, heteroscedasticity case}

\item{etype}{type of error, where 
etype = 1 for N(0,1),
etype = 2 for t_3, 
etype = 3 for 0.9N(0,1) + 0.1t_3, 
etype = 4 for 0.9N(0,1) + 0.1Cauchy(0,1).}
}
\value{
A matrix with the elements
\item{y}{The response variable.}
\item{x}{The scalar covariate with a change point}
\item{z}{A vector of covariates.}
}
\description{
The function generates data from the bent line quantile regression model.
}
\details{
The bent line quantile regression model: 
 Q_tau (Y|x,z) = [1, x, (x-t_0)_+, z] * bet0.
}
\examples{
\dontrun{

## simulated data
n <- 200
t0 <- 1.5    # true change point
bet0 <- c(1, 3, -2, 1)     # true regression coefficients
tau <- 0.7
modtype <- 1
etype <- 1
dat <- bentQRsimdat(n, bet0, t0, tau, modtype, etype)
head(dat)
}

}
\author{
Feipeng Zhang
}
\keyword{bentQRsimdat}

